+++ src/services/email.service.ts
@@ -2,12 +2,11 @@ import SMTPTransport from "nodemailer/lib/smtp-transport";
 import nodemailer from "nodemailer";
 import { IEmailRecipient } from "../models/email.models";
 import { SYSTEM_EMAIL_CREDENTIALS } from "../constants";
-import { NOREPLY_EMAIL } from "./../constants/index";
-import { getDictionayByLanguage } from "./../utils/localization/index";
+import { NOREPLY_EMAIL } from "./../constants";
 import log from "../utils/logs";
 
 export const sendEmailToUser = async (userRecipient: IEmailRecipient) => {
-  const { name, email, subject, message } = userRecipient;
+  const { name, email, subject, messageHTML, messagePlainText } = userRecipient;
 
   let transporter = nodemailer.createTransport({
     host: "smtp.office365.com",
@@ -26,15 +25,15 @@ export const sendEmailToUser = async (userRecipient: IEmailRecipient) => {
       to: `"${name}" <${email}>`, // list of receivers
       subject: subject, // Subject line
       replyTo: email,
-      text: message, // plain text body
-      html: message,
+      text: messagePlainText, // plain text body
+      html: messageHTML,
     })
     .then((emailInfo: SMTPTransport.SentMessageInfo) => {
       log.success(`Email successfuly sent to ${email}`);
       return emailInfo;
     })
-    .catch(() => {
-      log.error(`Error sending email to user ${email}`);
+    .catch((error: any) => {
+      log.error(`Error sending email to user ${email}:`, error);
       return null;
     });
 };