+++ src/services/webhooks/invoice.service.ts
@@ -2,12 +2,14 @@ import { IEmailRecipient } from "../../models/email.models";
 import { IInvoice } from "../../models/invoices.models";
 import { getDictionayByLanguage } from "../../utils/localization";
 import log from "../../utils/logs";
-import {
-  convertPaymentAmountToDecimalString,
-  getLanguageFromCurrency,
-} from "../../utils/stripe";
+import { getLanguageFromCurrency } from "../../utils/stripe";
 import { sendEmailToUser } from "../email.service";
 import UserDb from "../../models/user.models";
+import {
+  getHTMLBody,
+  getHTMLButton,
+  getHTMLFooterByLanguage,
+} from "../../utils/email";
 
 export const handleInvoice = async (event: any) => {
   const invoice: IInvoice = event.data.object as IInvoice;
@@ -21,48 +23,65 @@ export const handleInvoice = async (event: any) => {
     const language = getLanguageFromCurrency(currency);
     const dictionary = getDictionayByLanguage(language);
     let userRecipient: IEmailRecipient | undefined = undefined;
-    const amount: string = convertPaymentAmountToDecimalString(
-      invoice.amount_paid
-    );
-
-    log.success(`📄 Nota fiscal ${userFound.email} disponível.`);
-
-    const getEmailMessageByLanguage = (lang: string) => {
-      switch (lang) {
-        case "pt":
-          return `
-          <b>Olá, ${userFound.firstName}!</b><br>
-          <br>
-          A nota fiscal para a sua nova assinatura já está disponível!<br>
-          Você pode visualizar e baixar aqui: <a href"${invoice.invoice_pdf}">CLICANDO AQUI</a><br>
-          <br>
-          <br>
-          Equipe Socialbio<br>
-          <a href"https://socialbio.me">https://www.socialbio.me</a>   
-          `;
-        default:
-          return `
-          <b>Hey ${userFound.firstName},</b><br>
-          <br>
-          The invoice for your new subscription is available!<br>
-          You can view and download <a href"${invoice.invoice_pdf}">CLICKING HERE</a><br>
-          <br>
-          <br>
-          Socialbio Team<br>
-          <a href"https://socialbio.me">https://www.socialbio.me</a>   
-          `;
-      }
-    };
 
     switch (event.type) {
-      case "invoice.paid":
+      case "invoice.paid": {
+        const getEmailMessageByLanguage = (lang: string) => {
+          switch (lang) {
+            case "pt":
+              return {
+                html: getHTMLBody(`
+              <b>Olá, ${userFound.firstName}!</b><br>
+              <br>
+              A nota fiscal para a sua nova assinatura já está disponível!<br>
+              Você pode visualizar e baixar aqui:
+              <br>
+              <br>
+              ${getHTMLButton(invoice.invoice_pdf, "NOTA FISCAL")}
+              <br>
+              <br>
+              ou abrindo esse link no seu navegador: ${invoice.invoice_pdf}.<br>
+              <br>
+              <br>
+              ${getHTMLFooterByLanguage(language)}
+              `),
+                text: `
+              Olá, ${userFound.firstName}! A nota fiscal para a sua nova assinatura já está disponível! Você pode visualizar e baixar abrindo esse link no seu navegador: ${invoice.invoice_pdf}. Equipe Socialbio (https://www.socialbio.me)`,
+              };
+            default:
+              return {
+                html: getHTMLBody(`
+              <b>Hey ${userFound.firstName},</b><br>
+              <br>
+              The invoice for your new subscription is available!<br>
+              You can view and download here:
+              <br>
+              <br>
+              ${getHTMLButton(invoice.invoice_pdf, "INVOCIE")}
+              <br>
+              <br>
+              or opening this link in your browser: ${invoice.invoice_pdf}<br>
+              <br>
+              <br>
+              ${getHTMLFooterByLanguage(language)}
+              `),
+                text: `
+              Hey ${userFound.firstName}! The invoice for your new subscription is available! You can view and download opening this link in your browser: ${invoice.invoice_pdf}. Socialbio Team (https://www.socialbio.me)`,
+              };
+          }
+        };
+
+        log.success(`📄 Nota fiscal ${userFound.email} disponível.`);
+
         userRecipient = {
           name: userFound.firstName,
           email: userFound.email,
           subject: `[Socialbio] ${dictionary.paymentSucceed}`,
-          message: getEmailMessageByLanguage(language),
+          messageHTML: getEmailMessageByLanguage(language).html,
+          messagePlainText: getEmailMessageByLanguage(language).text,
         };
         break;
+      }
       default:
         log.error(`Unhandled event type ${event.type}`);
         break;