+++ src/services/webhooks/payment-intent.service.ts
@@ -14,6 +14,7 @@ import SubscriptionsDB, {
 import log from "../../utils/logs";
 import { SYSTEM_EMAIL_CREDENTIALS } from "../../constants";
 import { sendEmailToUser } from "../email.service";
+import { getHTMLBody } from "../../utils/email";
 
 export const updateSubscriptionFromPaymentIntent = async (
   paymentIntent: IPaymentIntent
@@ -63,26 +64,54 @@ export const handlePaymentIntent = async (event: any) => {
       const plan = getPlanByPriceId(updatedSubscription.priceId);
 
       switch (event.type) {
-        case "payment_intent.payment_failed":
+        case "payment_intent.payment_failed": {
           log.warn(`⛈️ Pagamento do usuario ${userFound.email} falhou`);
+
+          const getEmailMessageByLanguage = (lang: string) => {
+            switch (lang) {
+              case "pt":
+                return {
+                  html: getHTMLBody(`
+                  <b>Olá, ${userFound.firstName}!</b><br>
+                  <br>
+                  Desculpe! Seu pagamento falhou!<br>
+                  Por favor tente fazer sua assinatura novamente 
+                  <a href="https://socialbio.me">CLICANDO AQUI</a> ou abrindo esse link no seu navegador: https://socialbio.me.
+                  <br>
+                  <br>
+                  Equipe Socialbio<br> 
+                  `),
+                  text: `
+                  Olá, ${userFound.firstName}! Parabéns, seu pagamento foi finalizado com sucesso! Agora você é um assinante ${plan}! Bem-vindo a bordo! Equipe Socialbio (https://www.socialbio.me)`,
+                };
+              default:
+                return {
+                  html: `
+                  <b>Hey ${userFound.firstName},</b><br>
+                  <br>
+                  Sorry, your payment failed!<br>
+                  Please try to subscribe again.<br>
+                  <a href="https://socialbio.me">Click here to try again</a> or open this link your browser: https://socialbio.me.
+                  <br>
+                  <br>
+                  Socialbio Team<br>  
+                  `,
+                  text: `
+                  Hey ${userFound.firstName}! Sorry, your payment failed! Please try to subscribe again. Try again opening this link in your browser: https://socialbio.me. Socialbio Team`,
+                };
+            }
+          };
 
           userRecipient = {
             name: userFound.firstName,
             email: userFound.email,
             subject: `[Socialbio] ${dictionary.payment}`,
-            message: `
-        <b>Hey ${userFound.firstName},</b><br>
-        <br>
-        Sorry, your payment failed!<br>
-        Please try to subscribe again.<br>
-        <a href"https://socialbio.me">Click here to try again.</a>        
-        <br>
-        <br>
-        Socialbio Team<br>  
-        `,
+            messageHTML: getEmailMessageByLanguage(language).html,
+            messagePlainText: getEmailMessageByLanguage(language).text,
           };
           break;
-        case "payment_intent.succeeded":
+        }
+        case "payment_intent.succeeded": {
           log.success(
             "💰 Pagamento com sucesso para o usuario " + userFound.email
           );
@@ -90,7 +119,8 @@ export const handlePaymentIntent = async (event: any) => {
           const getEmailMessageByLanguage = (lang: string) => {
             switch (lang) {
               case "pt":
-                return `
+                return {
+                  html: getHTMLBody(`
               <b>Olá, ${userFound.firstName}!</b><br>
               <br>
               Parabéns! Seu pagamento foi finalizado com sucesso!<br>
@@ -99,10 +129,15 @@ export const handlePaymentIntent = async (event: any) => {
               <br>
               <br>
               Equipe Socialbio<br>
-              <a href"https://socialbio.me">https://www.socialbio.me</a>   
-              `;
+              <a href="https://socialbio.me">https://www.socialbio.me</a>   
+              `),
+                  text: `
+              Olá, ${userFound.firstName}! Parabéns, seu pagamento foi finalizado com sucesso! Agora você é um assinante ${plan}! Bem-vindo a bordo! Equipe Socialbio (https://www.socialbio.me)
+              `,
+                };
               default:
-                return `
+                return {
+                  html: getHTMLBody(`
               <b>Hey ${userFound.firstName},</b><br>
               <br>
               Congratulations! Your payment was finished with success!<br>
@@ -113,8 +148,16 @@ export const handlePaymentIntent = async (event: any) => {
               <br>
               <br>
               Socialbio Team<br>
-              <a href"https://socialbio.me">https://www.socialbio.me</a>   
-              `;
+              <a href="https://socialbio.me">https://www.socialbio.me</a>   
+              `),
+                  text: `
+              Hey ${
+                userFound.firstName
+              }! Congratulations, your payment was finished with success! You are now a ${getPlanByPriceId(
+                    updatedSubscription.priceId
+                  )} subscriber! Welcome onboard! Socialbio Team (https://www.socialbio.me)
+              `,
+                };
             }
           };
 
@@ -122,7 +165,8 @@ export const handlePaymentIntent = async (event: any) => {
             name: userFound.firstName,
             email: userFound.email,
             subject: `[Socialbio] ${dictionary.paymentSucceed}`,
-            message: getEmailMessageByLanguage(language),
+            messageHTML: getEmailMessageByLanguage(language).html,
+            messagePlainText: getEmailMessageByLanguage(language).text,
           };
 
           if (SYSTEM_EMAIL_CREDENTIALS?.user) {
@@ -130,18 +174,24 @@ export const handlePaymentIntent = async (event: any) => {
               name: "System",
               email: SYSTEM_EMAIL_CREDENTIALS.user,
               subject: `[Socialbio] ${dictionary.payment}`,
-              message: `
+              messageHTML: getHTMLBody(`
               <b>Hey Team!</b><br>
               <br>
               User ${userFound.firstName} (${
                 userFound.email
-              }) has just paid for a ${plan} subscription of ${amount} (${currency.toUpperCase()})<br>
+              }) has just paid for a ${plan} subscription of ${amount} (${currency.toUpperCase()})
+                <br>
               Socialbio System<br>  
-              `,
+              `),
+              messagePlainText: `
+              Hey Team! User ${userFound.firstName} (${
+                userFound.email
+              }) has just paid for a ${plan} subscription of ${amount} (${currency.toUpperCase()}). Socialbio System`,
             };
             sendEmailToUser(systemRecipient);
           }
           break;
+        }
         default:
           console.log(`Unhandled event type ${event.type}`);
           break;